Fix #111: Add error/failure logging and centralize package configs
Logging:
We want to to log errors and failures by default to prevent those
problems from being swallowed up since they may contain important
information about underlying issues.

EPackageConfigurations:
Centralize EPackageConfiguration management and apply it on the Java
model server client to ensure proper setup. We also ensure that the
ChangePackage is properly initialized on the client-side. We already use
the RecordingModelResourceManager as default on the server so we need
the ChangePackage to be initialized.