#140 Update to emfcloud.checkstyle 0.1.0-RC02
The new release candiate uses a ruleset that is compliant with Checkstyle 8.44 as default.
This ensures that the checkstyle config can also be used in the latest Eclipse Relaease (8.44).
If you don't use the latest Eclipse release it's still recommended to update to the latest eclipse-checkstyle plugin version. However, the old checkstyle config (8.39) is still available and can be set in the project-specific properties.

Fixes #140