Fix #174 API v2: Injection of JsonPatchHelper
Inject the JsonPatchHelper also into the resource manager and
transaction controller that previously constructed it for themselves.
This requires additionally:
- provider injection for the resource manager because the
  patch helper depends on the resource manager
- configuration of injectors in tests now needs to include injection
  of codecs map because it is injected into the patch helper

So far, this is tied to json-v2 format exclusively as this is the
default format for API v2. If the helper should need to pick up
other formats from the request/socket context, then that's an
orthogonal issue.

Contributed on behalf of STMicroelectronics.

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>