#205: API v2: Subscription option for URI fragments in update notifications
Add new subscription option to the V2 client API and also to the
example CLI app.

As the number of subscription options is growing, it seems
appropriate to add a more generic API to capture them all together
in one bundle in one subscription call without further proliferating
overloads of the subscribe method signature.

Contributed on behalf of STMicroelectronics.

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>